requirements.txt
arxiv>=2.1.0
feedparser>=6.0
httpx>=0.27.0
jinja2>=3.1.0
pyyaml>=6.0
